backend/requirements.txt
# Django Core
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment Management
python-dotenv==1.0.0

# Async & WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1

# CORS
django-cors-headers==4.3.1

# Utilities
gunicorn==21.2.0

# Binance Scanner Dependencies
aiohttp==3.9.1
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
scipy==1.11.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
